Move `ExerciseListViewController` to production
