+++ Gymondo/GymondoiOSTests/ExerciseListViewControllerTests.swift
@@ -8,30 +8,7 @@
 import XCTest
 import UIKit
 @testable import Gymondo
-
-final class ExerciseListViewController: UITableViewController {
-    private var viewModel: ExerciseListViewModelLogic?
-
-    convenience init(viewModel: ExerciseListViewModelLogic) {
-        self.init()
-        self.viewModel = viewModel
-    }
-
-    override func viewDidLoad() {
-        super.viewDidLoad()
-        refreshControl = UIRefreshControl()
-        refreshControl?.addTarget(self, action: #selector(refresh), for: .valueChanged)
-        refresh()
-    }
-
-    @objc private func refresh() {
-        refreshControl?.beginRefreshing()
-        viewModel?.start { [weak self] in
-            self?.refreshControl?.endRefreshing()
-        }
-    }
-
-}
+@testable import GymondoiOS
 
 final class ExerciseListViewControllerTests: XCTestCase {
